% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoter_strength.R
\name{DIPA}
\alias{DIPA}
\title{Dynamic promoter activity.}
\usage{
DIPA(x, t.f, f.data, t.b, b.data, m = log(2)/0.45, d = log(2)/0.5,
  beta = 1, d.r = 1, data.type = 3)
}
\arguments{
\item{x}{Vector of times over which to calculate promoter activity (in hours).}

\item{t.f}{Vector of times of reporter protein (or mRNA) measurements (in hours).}

\item{f.data}{Vector of of reporter protein (or mRNA) measurements.}

\item{t.b}{Vector of times of biomass measurements (in hours).}

\item{b.data}{Vector of biomass measurements.}

\item{m}{Maturation constant (if using fluorescent protein reporter).
Defaults to log(2)/0.45 h^-1.}

\item{d}{Degradation rate of protein. Defaults to log(2)/0.5 h^-1 (Mateus & Avery
2000).}

\item{beta}{Translation rate. Defaults to 1.}

\item{d.r}{mRNA degradation rate. Defaults to 1.}

\item{data.type}{Indicates type of input data, with numbers corresponding to
number of ODEs necessary to model the reporter system. 1 = mRNA, 2 =
protein other than fluorescent (i.e. luciferase, beta-galactosidase, etc.),
3 = fluorescent protein (i.e. GFP, RFP, etc). For more information, see
references. Defaults to 3.}
}
\value{
Returns a list \code{L} with \code{L$t} the time series input by the
  user and \code{L$promact} the calculated promoter activities at these
  times.
}
\description{
\code{DIPA} returns a vector of promoter activity over specified
times as described in PUBLICATION NAME.
}
\details{
This function requires reporter protein data and corresponding time points,
and biomass data and corresponding time points stored in R vectors or similar
data structure. User can choose the nature of input (fluorescent protein,
other protein or mRNA) data via the data.type input. For single-cell data,
input biomass as a vector of 1's of corresponding length to gene expression
measurements, and consider promoter activity output to be activity per cell.
}
\section{Author}{

Soumya Kannan, Technical University of Denmark, 2017.
}

\references{
PUBLICATION CITATION.
}
