% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_1}
\alias{sim_1}
\title{Constant fluorescence during stationary phase}
\format{A data frame with 30100 rows and 4 variables:
\describe{
   \item{sim_num}{simulation number (1--100)}
   \item{time}{time in hours (0--30)}
   \item{biomass}{simulated biomass measurement ()}
   \item{fluorescence}{simulated fluorescence measurement()}
}}
\usage{
sim_1
}
\description{
A simulation of biomass and fluorescence data with fluorescence observed 
during stationary phase. Variables are as follows:
}
\keyword{datasets}
