% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_3}
\alias{sim_3}
\title{Constant fluorescence during lag phase}
\format{A data frame with 30100 rows and 4 variables:
\describe{
   \item{sim_num}{simulation number (1--100)}
   \item{time}{time in hours (0--30)}
   \item{biomass}{simulated biomass measurement ()}
   \item{fluorescence}{simulated fluorescence measurement()}
}}
\usage{
sim_3
}
\description{
A simulation of biomass and fluorescence with constant fluorescence in lag phase 
that decreases exponentially during exponential phase. Variables are as follows:
}
\keyword{datasets}
