% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{step_size_sims}
\alias{step_size_sims}
\title{Step function of promoter activity with different data step sizes}
\format{A data frame with 41400 rows and 5 variables:
\describe{
   \item{step_size}{Step size/distance between data points in time (0.1, 0.5, 1, 1.5)}
   \item{sim_num}{simulation number (1--100)}
   \item{time}{time in hours (0--30)}
   \item{biomass}{simulated biomass measurement ()}
   \item{fluorescence}{simulated fluorescence measurement()}
}}
\usage{
step_size_sims
}
\description{
Simulations of biomass and fluorescence data for a step function of promoter activity
occurring from 12-16 h. Includes 100 simulations each for 4 different step sizes of 
data. Variables are as follows
}
\keyword{datasets}
